services/ml-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pandas==2.2.2
numpy==1.26.4
psycopg2-binary==2.9.9
scikit-learn==1.5.1
joblib==1.4.2
# torch==2.3.1